Fixed last error for creating the Students Api
