+++ build.gradle
@@ -35,7 +35,5 @@ dependencies {
     compile group: 'com.amazonaws', name: 'aws-lambda-java-events', version: '2.2.7'
     compile group: 'org.postgresql', name: 'postgresql', version: '42.1.4'
 
-    compileOnly group: 'org.projectlombok', name: 'lombok', version: '1.18.12'
-
     testCompile group: 'junit', name: 'junit', version: '4.12'
 }
