+++ src/main/java/br/com/studiesMaterials/handlers/BookHandler.java
@@ -27,6 +27,7 @@ public class BookHandler implements BookDao {
             );
 
             statement.executeUpdate(sql);
+            conn.close();
 
             responseEvent.setStatusCode(201);
 
@@ -81,6 +82,7 @@ public class BookHandler implements BookDao {
             String sql = String.format("DELETE FROM public.book WHERE id = '%s'", bookId);
 
             statement.executeUpdate(sql);
+            conn.close();
 
             responseEvent.setStatusCode(204);
 
@@ -88,9 +90,6 @@ public class BookHandler implements BookDao {
         } catch (SQLException error) {
             error.printStackTrace();
             responseEvent.setStatusCode(500);
-
-            String body = String.format("{ 'message': %s}", error.getMessage());
-            responseEvent.setBody(body);
 
             return responseEvent;
         }
