+++ src/main/java/br/com/studiesMaterials/web/api/responses/StudentResponse.java
@@ -1,0 +1,13 @@
+package br.com.studiesMaterials.web.api.responses;
+
+import java.util.UUID;
+
+public class StudentResponse {
+    UUID id;
+    String username;
+
+    public StudentResponse(String id,String username) {
+        this.id = UUID.fromString(id);
+        this.username = username;
+    }
+}
