--- src/main/java/br/com/studiesMaterials/web/api/schemas/CourseResponse.java
@@ -1,17 +1,0 @@
-package br.com.studiesMaterials.web.api.schemas;
-
-import java.util.UUID;
-
-public class CourseResponse {
-    private UUID id;
-    private String name;
-    private String platform;
-    private Float price;
-
-    public CourseResponse(String id, String name, String platform, String price){
-        this.id = UUID.fromString(id);
-        this.name = name;
-        this.platform = platform;
-        this.price = Float.parseFloat(price);
-    }
-}
