--- src/main/java/br/com/studiesMaterials/web/api/schemas/PodcastResponse.java
@@ -1,20 +1,0 @@
-package br.com.studiesMaterials.web.api.schemas;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class PodcastResponse {
-    public String id;
-    public String subject;
-    public String time;
-    public String link;
-
-    public PodcastResponse(String id, String subject, String time, String link) {
-        this.id = id;
-        this.subject = subject;
-        this.time = time;
-        this.link = link;
-    }
-}
