+++ src/main/java/br/com/studiesMaterials/web/api/schemas/PodcastSchema.java
@@ -1,10 +1,5 @@
 package br.com.studiesMaterials.web.api.schemas;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
 public class PodcastSchema {
     public String subject;
     public int time;
