--- src/main/java/br/com/studiesMaterials/web/api/schemas/StudentResponse.java
@@ -1,11 +1,0 @@
-package br.com.studiesMaterials.web.api.schemas;
-
-public class StudentResponse {
-    public String id;
-    public String username;
-
-    public StudentResponse(String id,String username) {
-        this.id = id;
-        this.username = username;
-    }
-}
